Extremely basic stop() support; gotoEnd is always true for the moment
This patch fails to do a couple things, including back-calculating
value by process at time of call, testing to see if we were even
running, or masking the complete message that we set a timer for
earlier in execution. It does, however, do the very small amount I need
it to do.